refactor qingcloud errors
