+++ cloud/services/networking/router.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	infrav1beta1 "github.com/kubesphere/cluster-api-provider-qingcloud/api/v1beta1"
-	utilerrors "github.com/kubesphere/cluster-api-provider-qingcloud/util/errors"
+	qcerrors "github.com/kubesphere/cluster-api-provider-qingcloud/util/errors/qingcloud"
 	qcs "github.com/yunify/qingcloud-sdk-go/service"
 )
 
@@ -26,7 +26,7 @@ func (s *Service) CreateRouter(securityGroupID string) (infrav1beta1.QCResourceI
 	}
 
 	if qcs.IntValue(o.RetCode) != 0 {
-		return nil, utilerrors.NewQingCloudError(o.RetCode, o.Message)
+		return nil, qcerrors.NewQingCloudError(o.RetCode, o.Message)
 	}
 
 	return o.Routers[0], nil
@@ -47,7 +47,7 @@ func (s *Service) GetRouter(routerID infrav1beta1.QCResourceID) (*qcs.DescribeRo
 		return nil, err
 	}
 	if qcs.IntValue(o.RetCode) != 0 {
-		return nil, utilerrors.NewQingCloudError(o.RetCode, o.Message)
+		return nil, qcerrors.NewQingCloudError(o.RetCode, o.Message)
 	}
 
 	return o, nil
@@ -67,7 +67,7 @@ func (s *Service) JoinRouter(routerID, vxnetID infrav1beta1.QCResourceID, ipNetw
 	}
 
 	if qcs.IntValue(o.RetCode) != 0 {
-		return utilerrors.NewQingCloudError(o.RetCode, o.Message)
+		return qcerrors.NewQingCloudError(o.RetCode, o.Message)
 	}
 
 	return nil
@@ -92,7 +92,7 @@ func (s *Service) UnbindingRouterVxnets(routerID infrav1beta1.QCResourceID, vxne
 		return err
 	}
 	if qcs.IntValue(l.RetCode) != 0 {
-		return utilerrors.NewQingCloudError(l.RetCode, l.Message)
+		return qcerrors.NewQingCloudError(l.RetCode, l.Message)
 	}
 
 	//a, err := c.UpdateRouters(
@@ -117,7 +117,7 @@ func (s *Service) DeleteRoute(routerID infrav1beta1.QCResourceID) error {
 		return err
 	}
 	if qcs.IntValue(d.RetCode) != 0 {
-		return utilerrors.NewQingCloudError(d.RetCode, d.Message)
+		return qcerrors.NewQingCloudError(d.RetCode, d.Message)
 	}
 	return nil
