+++ cloud/services/networking/securitygroup.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	infrav1beta1 "github.com/kubesphere/cluster-api-provider-qingcloud/api/v1beta1"
-	utilerrors "github.com/kubesphere/cluster-api-provider-qingcloud/util/errors"
+	qcerrors "github.com/kubesphere/cluster-api-provider-qingcloud/util/errors/qingcloud"
 	qcs "github.com/yunify/qingcloud-sdk-go/service"
 )
 
@@ -24,7 +24,7 @@ func (s *Service) CreateSecurityGroup() (infrav1beta1.QCResourceID, error) {
 	}
 
 	if qcs.IntValue(o.RetCode) != 0 {
-		return nil, utilerrors.NewQingCloudError(o.RetCode, o.Message)
+		return nil, qcerrors.NewQingCloudError(o.RetCode, o.Message)
 	}
 
 	return o.SecurityGroupID, nil
@@ -43,7 +43,7 @@ func (s *Service) AddSecurityGroupRules(securityGroupID infrav1beta1.QCResourceI
 		return nil, err
 	}
 	if qcs.IntValue(o.RetCode) != 0 {
-		return nil, utilerrors.NewQingCloudError(o.RetCode, o.Message)
+		return nil, qcerrors.NewQingCloudError(o.RetCode, o.Message)
 	}
 
 	a, err := c.ApplySecurityGroup(
@@ -55,7 +55,7 @@ func (s *Service) AddSecurityGroupRules(securityGroupID infrav1beta1.QCResourceI
 		return nil, err
 	}
 	if qcs.IntValue(a.RetCode) != 0 {
-		return nil, utilerrors.NewQingCloudError(a.RetCode, a.Message)
+		return nil, qcerrors.NewQingCloudError(a.RetCode, a.Message)
 	}
 
 	return o.SecurityGroupRules[0], nil
@@ -76,7 +76,7 @@ func (s *Service) GetSecurityGroup(securityGroupID infrav1beta1.QCResourceID) (*
 		return nil, err
 	}
 	if qcs.IntValue(o.RetCode) != 0 {
-		return nil, utilerrors.NewQingCloudError(o.RetCode, o.Message)
+		return nil, qcerrors.NewQingCloudError(o.RetCode, o.Message)
 	}
 
 	return o, nil
@@ -96,7 +96,7 @@ func (s *Service) DeleteSecurityGroup(securityGroupID infrav1beta1.QCResourceID)
 	}
 
 	if qcs.IntValue(o.RetCode) != 0 {
-		return utilerrors.NewQingCloudError(o.RetCode, o.Message)
+		return qcerrors.NewQingCloudError(o.RetCode, o.Message)
 	}
 
 	return nil
