+++ cloud/services/networking/vxnet.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	infrav1beta1 "github.com/kubesphere/cluster-api-provider-qingcloud/api/v1beta1"
-	utilerrors "github.com/kubesphere/cluster-api-provider-qingcloud/util/errors"
+	qcerrors "github.com/kubesphere/cluster-api-provider-qingcloud/util/errors/qingcloud"
 	qcs "github.com/yunify/qingcloud-sdk-go/service"
 )
 
@@ -25,7 +25,7 @@ func (s *Service) CreateVxNet() (infrav1beta1.QCResourceID, error) {
 	}
 
 	if qcs.IntValue(o.RetCode) != 0 {
-		return nil, utilerrors.NewQingCloudError(o.RetCode, o.Message)
+		return nil, qcerrors.NewQingCloudError(o.RetCode, o.Message)
 	}
 
 	return o.VxNets[0], nil
@@ -46,7 +46,7 @@ func (s *Service) GetVxNet(vxnetID infrav1beta1.QCResourceID) (*qcs.DescribeVxNe
 	}
 
 	if qcs.IntValue(o.RetCode) != 0 {
-		return nil, utilerrors.NewQingCloudError(o.RetCode, o.Message)
+		return nil, qcerrors.NewQingCloudError(o.RetCode, o.Message)
 	}
 
 	return o, nil
@@ -66,7 +66,7 @@ func (s *Service) DeleteVxNet(vxnetID infrav1beta1.QCResourceID) error {
 	}
 
 	if qcs.IntValue(o.RetCode) != 0 {
-		return utilerrors.NewQingCloudError(o.RetCode, o.Message)
+		return qcerrors.NewQingCloudError(o.RetCode, o.Message)
 	}
 
 	return nil
@@ -87,7 +87,7 @@ func (s *Service) LeaveRouter(routerID, vxnetID infrav1beta1.QCResourceID) error
 	}
 
 	if qcs.IntValue(l.RetCode) != 0 {
-		return utilerrors.NewQingCloudError(l.RetCode, l.Message)
+		return qcerrors.NewQingCloudError(l.RetCode, l.Message)
 	}
 
 	return nil
